Disable IPv6 when building curl on macOS
When `ENABLE_IPV6` is ON, libcurl will depend on the frameworks on
macOS. Then users must add the following link options when linking the
static curl library.

```
-framework CoreFoundation -framework CoreServices -framework SystemConfiguration
```

It's not user-friendly if users do not need the IPv6 feature.

In pulsar-client-cpp's previous build process, the `--disable-ipv6`
option is also added to disable this feature.